Move htrun dependency out of mbed-os-tools
